backend/requirements.txt
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
sqlparse==0.4.4
rake_nltk==1.0.6
groq>=0.28
gunicorn==21.2.0
